:truck: (901): move split_audio_on_timestamps and convert_to_french_speaker to domain
