--- mcr-core/mcr_meeting/app/services/speech_to_text/utils/__init__.py
@@ -1,7 +1,0 @@
-"""Public exports for the speech_to_text utils package."""
-
-from .audio import split_audio_on_timestamps
-
-__all__ = [
-    "split_audio_on_timestamps",
-]
